Rename index.js in Pages folder to homepage.js
